Fix cross-thread free stranding and the broadcast herd; add the workloads
A reader thread allocating big buffers that a pool of workers frees
regressed badly in the last round (measured with a video pipeline of
690 KiB frames and forty workers: 1.4 GB resident, 370k page faults,
twice the time per frame). Three causes, all fixed:

- Remote frees were batched per freeing thread and only flushed after
  64 entries, so a worker that freed one frame per work item stranded
  up to 63 of them. Batches now flush at 32 KiB of pending blocks and
  whenever the thread's caches are refilled or flushed, and the owner
  sweeps all its spans for remote frees every 250 ms so memory freed
  for it comes back even for classes it stopped allocating.
- The mutex stopped spinning as soon as anybody slept on it, turning
  short hand-offs into sleep/wake pairs; it now spins ten iterations,
  and a wake-in-flight bit keeps unlocks from waking another sleeper
  while the previous one is still on its way (half the futex calls of
  the 40-worker pipeline). Condition variable broadcasts wake one
  waiter and requeue the rest onto the mutex instead of waking all of
  them to fight for it; timed waiters use a separate word. pthread_cond_t
  is 32 bytes.
- Per-thread caches are back to 64 KiB per class and the empty-span
  reserve to 16 MiB per thread, so forty threads do not pin hundreds of
  megabytes.

The remote-free path is no longer compiled as cold, and the benchmark
suite gains fan-out free, parallel free and frame pipeline workloads
(all now led by rustlibc), plus a 16-thread and a 40-worker variant.

+++ rust-libc/include/pthread.h
@@ -16,7 +16,7 @@ typedef int pthread_key_t;
 typedef int pthread_once_t;
 typedef int pthread_spinlock_t;
 typedef struct { unsigned __s[5]; } pthread_mutex_t;
-typedef struct { unsigned __s[4]; } pthread_cond_t;
+typedef struct { unsigned long __s[4]; } pthread_cond_t;
 typedef struct { unsigned __s[4]; } pthread_rwlock_t;
 typedef struct { unsigned long __s[2]; } pthread_barrier_t;
 typedef struct { unsigned __type; } pthread_mutexattr_t;
